Fixed get_relation_types; minor code refactor.
